🎁 Add conditional for Mobius work types
This commit will add a few decorators.  First, we have the
Site::Decorator which allows us to conditionally show/hide work types
based on what tenant we're on.  That brings us to our second decorator.
The AccountDecorator holds the array of tenant names that are under the
Mobius umbrella.  We add a helper method to check if our current tenant
is apart of Mobius.  Finally we tweak the view to hide the work types
based on the tenant.