🎁 Ensure CsvCollectionEntries are created
Prior, Mobius collections were coming in as CsvEntry objects and not
CsvCollectionEntry objects.  This was fixed with the CsvParserDecorator
by adding the model key and setting it to `'CollectionResource'`.  There
was a change in Bulkrax to account for this as well so we updated the
submodule's Bulkrax SHA as well. Lastly, there was some string
manipulation added to HasLocalProcessing as well.

Ref:
  - https://github.com/scientist-softserv/hykuup_knapsack/issues/169