🎁 Add MOBIUS specific logic for files
This commit will add MOBIUS specific logic.  MOBIUS works come in as a
tar.gz file with multiple CSVs and multiple folders of files.  The added
logic will combine the CSVs into one and move the files into a single
single directory called `files`.  This will allow Bulkrax to easily
process the files.  The combined CSVs will have a new `file` and `model`
column to allow Bulkrax to know which file to attach to which work.

Ref:
  - https://github.com/scientist-softserv/hykuup_knapsack/issues/170