🐛 Rescue RuntimeError
We're noticing that some of the MOBIUS PDFs were throwing errors when
trying to do the Hyrax text extraction.  Putting in a rescue for the
time being to prevent the error to stop thumbnail the derivatives
process.  Ultimately when we get IIIF Print Valkyrie ready, the OCR
would be done there instead.