🧹 Add logic to check for csv filename as well
This commit will add the logic to check that the CSV's filename ends
with `C&C.csv` as a fallback for marking rows as collections.  It was
noted that there are data decrpencies in the CSVs where not all
collections were marked as such.  But the client confirmed that all the
entries in the CSVs that end with C&C.csv are collections.